"Updated schema definitions for various tables, including AuthorCommi…

+++ lib/phx_book_store/orders_items/order_items.ex
@@ -7,7 +7,7 @@ defmodule PhxBookStore.OrdersItems.OrderItems do
 
   schema "order_items" do
     field :quantity, :integer
-    field :price, :float
+    field :price, :integer
     has_many :books, Book
     has_many :book_stores,BookStore
